refactor ContentArea component to utilize CreateContent for content creation; move tag input and media embedding functionality to CreateContent component for better separation of concerns
